audio: ril: ensure open RIL-connection when leaving ril_open()
This is required for the WB-AMR callback to be recevied on-time.

The RIL issues this callback before any other RIL-method has the
chance to open the connection the the RIL-daemon.

Furthermore, the RIL socket is being created in a rather late boot
stage, thus the need for a indefinite loop to ensure an open connection.

Change-Id: Ibca92067d7bf9d7c7d20f949d577295a3af669a9